refactor(database): update folder structure
Represents `db/collection/action.ts` structure
